// apps/web/src/lib/api/generated/model/invoiceResponseStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits
 * OpenAPI spec version: 1.0.0
 */

export type InvoiceResponseStatus =
  (typeof InvoiceResponseStatus)[keyof typeof InvoiceResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceResponseStatus = {
  OPEN: 'OPEN',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
} as const;