// apps/web/src/lib/api/generated/model/invoiceItemResponseSourceType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, billing, products
 * OpenAPI spec version: 1.0.0
 */

export type InvoiceItemResponseSourceType =
  (typeof InvoiceItemResponseSourceType)[keyof typeof InvoiceItemResponseSourceType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceItemResponseSourceType = {
  VISIT_FEE: 'VISIT_FEE',
  DISEASE: 'DISEASE',
  MEDICATION: 'MEDICATION',
  PRODUCT: 'PRODUCT',
  MISC: 'MISC',
} as const;