// apps/web/src/lib/api/generated/model/invoiceResponsePaymentMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, billing, products
 * OpenAPI spec version: 1.0.0
 */

export type InvoiceResponsePaymentMethod =
  (typeof InvoiceResponsePaymentMethod)[keyof typeof InvoiceResponsePaymentMethod];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceResponsePaymentMethod = {
  CASH: 'CASH',
  CARD: 'CARD',
  TRANSFER: 'TRANSFER',
} as const;